enhancement: support skipping kernel module logic in Module via skipKernelModule flag
This commit introduces an enhancement proposal for KMM to support a new optional boolean flag, `skipKernelModule`, in the Module custom resource.
This feature allows third-party operators to use KMM’s device plugin deployment functionality without loading or managing out-of-tree kernel modules, which is useful when the required kernel modules are already in-tree.